Moved functions defined inline in source files to a common utility header.
